added check for stale users on deletion

+++ lib/friendings/amahi_friending_api.rb
@@ -138,8 +138,17 @@ class AmahiFriendingApi
 		!(username.blank? or (username.length < 3 or username.length > 32) or !(username =~ /\A[A-Za-z][A-Za-z0-9]+\z/))
 	end
 
-	def self.delete_user(user_id, email)
+	def self.delete_user(user_id, email, type)
 		url = "#{BASE_URL}/user/#{user_id}"
+
+		if type == "stale"
+			unless email.blank?
+				user = User.where({remote_user: email}).first
+				user.delete unless user.blank?
+			end
+			return "success", {}
+		end
+
 		begin
 			response = RestClient.delete(url, headers={"Api-Key" => APIKEY, "content-type" => :json})
 			json = JSON.parse(response)
